ch(code-clean-up): Reorganize the codebase and update README
This commit reorganise the codebase and updates the README
